requirements.txt
opencv-python
matplotlib
numpy
scipy
pytesseract
google-cloud-vision
setuptools 



